add zod validation + fixes

+++ packages/commonwealth/server/routes/topics/delete_topic_handler.ts
@@ -1,6 +1,12 @@
+import { AppError } from '../../../../common-common/src/errors';
 import { TopicAttributes } from '../../models/topic';
 import { ServerControllers } from '../../routing/router';
 import { TypedRequest, TypedResponse, success } from '../../types';
+import z from 'zod';
+
+const Errors = {
+  ValidationError: 'Validation error',
+};
 
 type DeleteTopicRequestParams = {
   topicId: string;
@@ -17,10 +23,18 @@ export const deleteTopicHandler = async (
   const { user, chain } = req;
   const { topicId } = req.params;
 
+  const validationSchema = z.coerce.number();
+  const validationResult = validationSchema.safeParse(topicId);
+  if (validationResult.success === false) {
+    throw new AppError(
+      `${Errors.ValidationError}: ${validationResult.error.message}`
+    );
+  }
+
   await controllers.topics.deleteTopic({
     user,
     chain,
-    topicId: parseInt(topicId, 10),
+    topicId: validationResult.data,
   });
 
   return success(res, null);
